Correcting the early callsite megamorphization
One insight is that usually code that performs an operation will only
use a single type param, else it will use the generic methods instead
of the specialized ones. Thus the benchmark was not doing the right
thing -- it was making the benchmark callsites themselves megamorphic
-- we only want the class callsites to become megamorphic, not the
benchmark ones, so we simulate real life code.