Closes #145
When typing `new ...` the typer would see the existance of the `.apply`
method and would try to call it. Yet, the newly introduced call would
break the rest of the compilation, as the pattern for constructing a
new class is:
```
  Apply(Select(New(...), <init>), args)
```
The actual crash for this occurs late in the pipeline, in explicitouter.