Closes #138 by creating bridges after miniboxing
There's the new `minibox-bridges` phase to do this. The reason
that this is necessary is because doing the copying after duplicators
could produce several bridges of the same method, which was incorrect.

There are still problems that may require name mangling, but I will file
those as a separate bug.