Work on #150: strict conditions on signature trans
This patch imposes stricter conditions on signature transformations,
thus making it impossible to break the object inheritance (and
overriding) model.

To some extent it reverts the bridges patch (currently just commenting
it out, since it will be necessary for fixing bugs #135 and #147):
https://github.com/miniboxing/miniboxing-plugin/commit/e13db8ca07eedb90f5d2794bb11d289c8e8b3f01

Also, this patch reverts #138 and the change in Duplicators to transform
method signatues.

The following commits will test the warnings generated when signatures
can't be transformed -- -P:minibox:warn should let users know about this.