Closes #178 for good
The problem lied in the fact that we were keeping the `FunctionX`
representation and using the function to miniboxed function bridges
in `components/runtime/src/miniboxing/runtime/MiniboxedFunctionBridge.scala`.

These, in turn, were compiled with the option to automatically replace the
function representation and redirect the `FunctionX.apply` calls to
`MiniboxedFunctionX` calls, esentially producing an infinite loop.

To fix this, the miniboxed function representation is not forced when
compiling the library and nor is the apply redirection. This should fix
bug #177 for good...