Closes #184: Repacking objects for duplication
Objects in local scopes are made up of the module class
(which is a class) and a getter:
```scala
  case object Apple extends Object with Product with Serializable {
    ...
  }
  @volatile var Apple$module: Apple.type = _
  case <stable> def Apple(): Apple.type = {
    Apple$module = new Apple.type();
    Apple$module
  }
```
When duplicating the expanded representation, all kinds of
bad things happen:
1) Double `sourceModule`s are created, one from the
`ensureCompanionModule` for the class and one for the `def Apple()`,
conflicting
2) Assuming the companion module isn't created, the backend crashes
because tq"case object Apple".sourceModule is NoSymbol and
q"def Apple(): Apple.type".moduleClass is NoSymbol

So, basically we fix these 2 issues. The solution used for 1) is
heavy-handed and only works in Scala 2.11. With complete control
over the compiler, one can probably do a lot better than this.