Closes #246 Array accessors to go through mbarray
This is a long-standing issue that I only now had time to diagnose.
The symptom is the fact that array updates and applies are not
rewritten as expected, for example:
```
  a(0) = a(1) // a.update(0, a.apply(1))
```
when `a` is known to be of type `Array[T]` where `T` is miniboxed.
Now, the reason this happens is because before, we only transformed
cases where the update and/or apply were interleaved with coercions.

But in the case above, no coercions are necessary, so the miniboxing
plugin leaves array acces as before. What can be done to fix this?
Forcefully rewrite array {apply,update} to mock versions that return
`@storage(<T's storage type>) T`. Then, `minibox-coerce` retypes
everything:
```
  mock_update(a, 0, mock_apply(a, 1))
  // mock_update takes @storage(Long) T
  // mick_apply returns @storage(Long) T
```
Thus, the code is okay, albeit using the mock apply/update. Then,
in `minibox-commit` we replace `mock_{apply,update}` by `mbarry_{}`,
allowing the specialized accessors to do their work.